requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0